add swagger, tests, improve code style

+++ auth_service/src/api/v1/roles.py
@@ -11,17 +11,6 @@ from services.helpers import admin_required, hash_password
 admin = Blueprint('admin', __name__, url_prefix='/admin')
 
 
-@admin.cli.command('createsuperuser')
-@click.argument('name')
-@click.argument('password')
-def create_superuser(name, password):
-    hash = hash_password(password)
-    superuser = User(login=name, password=hash, is_admin=True)
-    db.session.add(superuser)
-    db.session.commit()
-    return 'Superuser created'
-
-
 @admin.route('roles/add', methods=['POST'])
 @jwt_required()
 @admin_required()
@@ -40,7 +29,7 @@ def create_role():
         ), HTTPStatus.BAD_REQUEST
     data = request.json
     name = data['name']
-    if Role.query.filter_by(name=name).first() is None:
+    if db.session.query(Role).filter_by(name=name).first() is None:
         try:
             new_role = Role(name=name, description=data.get('description'))
             db.session.add(new_role)
@@ -48,14 +37,14 @@ def create_role():
             db.rollback()
         else:
             db.session.commit()
-        return jsonify({'message': f'Role {name} created'}), HTTPStatus.CREATED
+        return jsonify({'message': 'Role created'}), HTTPStatus.CREATED
     return jsonify(
         message="Wrong data",
         errors=[{"name": f"Role name <{name}> already exists"}],
     ), HTTPStatus.BAD_REQUEST
 
 
-@admin.route('roles/get', methods=['GET'])
+@admin.route('roles/', methods=['GET'])
 @jwt_required()
 @admin_required()
 def roles_list():
@@ -93,54 +82,51 @@ def edit_role(role_uuid):
                 'message': 'The name field cannot be empty'
             }
         ), HTTPStatus.BAD_REQUEST
-    role = Role.query.get(role_uuid)
+    role = db.session.query(Role).get(role_uuid)
     if role is None:
         return jsonify(
             {
                 'message': 'Role with such uuid is not found.'
             }
         ), HTTPStatus.NOT_FOUND
-    Role.query.filter_by(id=role_uuid).update(data)
+    db.session.query(Role).filter_by(id=role_uuid).update(data)
     db.session.commit()
-    return jsonify(
-        message=f'Role with uuid <{role_uuid}> is edited'
-    ), HTTPStatus.OK
+    return jsonify(message='Role edited'), HTTPStatus.OK
 
 
 @admin.route('roles/<uuid:role_uuid>/delete', methods=['DELETE'])
 @jwt_required()
 @admin_required()
 def delete_role(role_uuid):
-    role = Role.query.filter_by(id=role_uuid).first()
+    role = db.session.query(Role).filter_by(id=role_uuid).first()
     if role is None:
         return jsonify(
             {
                 'message': 'Role with such uuid is not found.'
             }
         ), HTTPStatus.NOT_FOUND
-    Role.query.filter_by(id=role_uuid).delete()
+    db.session.query(Role).filter_by(id=role_uuid).delete()
     db.session.commit()
-    return jsonify(
-        message=f'Role with uuid <{role_uuid}> is deleted'
-    ), HTTPStatus.OK
+    return jsonify(message='Role deleted'), HTTPStatus.NO_CONTENT
 
 
 @admin.route("/users/<uuid:user_uuid>/roles", methods=["GET"])
 @jwt_required()
 @admin_required()
 def get_user_roles_list(user_uuid):
-    user = User.query.get(user_uuid)
+    user = db.session.query(User).get(user_uuid)
     if not user:
         return jsonify(
             {
                 'message': 'User with such uuid is not found.'
             }
         ), HTTPStatus.NOT_FOUND
-    user_roles = UserRole.query.filter_by(user_id=user_uuid).all()
+    user_roles = db.session.query(UserRole).filter_by(user_id=user_uuid).all()
     role_ids = [ro.role_id for ro in user_roles]
     result = []
     for role_id in role_ids:
-        one_role = Role.query.filter(Role.id == role_id).first().name
+        one_role = db.session.query(Role).filter(
+            Role.id == role_id).first().name
         result.append({'name': one_role})
     return jsonify(roles=result), HTTPStatus.OK
 
@@ -150,8 +136,8 @@ def get_user_roles_list(user_uuid):
 @admin_required()
 def add_user_role(user_uuid):
     data = request.get_json()
-    user = User.query.filter_by(id=user_uuid).first()
-    role = Role.query.filter_by(name=data['name']).first()
+    user = db.session.query(User).filter_by(id=user_uuid).first()
+    role = db.session.query(Role).filter_by(name=data['name']).first()
     if role is None:
         return jsonify(
             {
@@ -164,14 +150,20 @@ def add_user_role(user_uuid):
     return jsonify(message='Role for User created'), HTTPStatus.CREATED
 
 
-@admin.route('/users/<uuid:user_uuid>/delete-role', methods=['POST'])
+@admin.route('/users/<uuid:user_uuid>/delete-role', methods=['DELETE'])
 @jwt_required()
 @admin_required()
 def delete_user_role(user_uuid):
     data = request.get_json()
-    user = User.query.filter_by(id=user_uuid).first()
-    role = Role.query.filter_by(name=data['name']).first()
-    user_role = UserRole.query.filter_by(
+    user = db.session.query(User).filter_by(id=user_uuid).first()
+    role = db.session.query(Role).filter_by(name=data['name']).first()
+    if role is None:
+        return jsonify(
+            {
+                'message': 'Role with such name is not found.'
+            }
+        ), HTTPStatus.NOT_FOUND
+    user_role = db.session.query(UserRole).filter_by(
         user_id=user.id, role_id=role.id).first()
     db.session.delete(user_role)
     db.session.commit()
